requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==3.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
ffmpeg-python==0.2.0
future==0.18.3
gunicorn==21.2.0
inflection==0.5.1
packaging==23.1
Pillow==10.0.1
psycopg2==2.9.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
